requirements.txt
mysql-connector-python==8.2.0
Flask==2.3.3
Flask-Session==0.5.0
pad4pi==1.1.15
rpi-lcd==1.0.0
RPi.GPIO==0.7.1 